Remove incompatible instructions in hw2_asm.S and compile it.
The references to get_cycles and get_instet in the assembly
will jump the code to another assembly file with RDCYCLE/RDCYCLEH
instructions, so comment all instructions related to it.

And by modifying the Makefile in csrc directory, you can update
hw2_asm.asmbin by using the command "make update".